refactor(remove _document.tsx) remove _document.tsx and integrate it into layout.tsx
- Updated the project to fully utilize App Router, eliminating the need for the pages folder.
